+++ AuthPermissions.SupportCode/ShardingServices/IAccessDatabaseInformation.cs
@@ -7,12 +7,12 @@ using StatusGeneric;
 namespace AuthPermissions.SupportCode.ShardingServices;
 
 /// <summary>
-/// This defines the CRUD methods to the shardingsettings.json which contains a list of <see cref="DatabaseInformation"/> 
+/// This defines the CRUD methods to the sharding settings file which contains a list of <see cref="DatabaseInformation"/> 
 /// </summary>
 public interface IAccessDatabaseInformation
 {
     /// <summary>
-    /// This will return a list of <see cref="DatabaseInformation"/> in the shardingsettings.json file in the application
+    /// This will return a list of <see cref="DatabaseInformation"/> in the sharding settings file in the application
     /// </summary>
     /// <returns>If no file, then returns an empty list</returns>
     List<DatabaseInformation> ReadShardingSettingsFile();
@@ -25,17 +25,17 @@ public interface IAccessDatabaseInformation
     DatabaseInformation GetDatabaseInformationByName(string databaseInfoName);
 
     /// <summary>
-    /// This adds a new <see cref="DatabaseInformation"/> to the list in the current shardingsettings.json file.
-    /// If there are no errors it will update the shardingsettings.json file in the application.
+    /// This adds a new <see cref="DatabaseInformation"/> to the list in the current sharding settings file.
+    /// If there are no errors it will update the sharding settings file in the application.
     /// </summary>
     /// <param databaseInfoName="databaseInfo"></param>
     /// <returns>status containing a success message, or errors</returns>
     IStatusGeneric AddDatabaseInfoToJsonFile(DatabaseInformation databaseInfo);
 
     /// <summary>
-    /// This updates a <see cref="DatabaseInformation"/> already in the shardingsettings.json file.
+    /// This updates a <see cref="DatabaseInformation"/> already in the sharding settings file.
     /// It uses the <see cref="DatabaseInformation.Name"/> in the provided in the <see cref="DatabaseInformation"/> parameter.
-    /// If there are no errors it will update the shardingsettings.json file in the application.
+    /// If there are no errors it will update the sharding settings file in the application.
     /// </summary>
     /// <param databaseInfoName="databaseInfo"></param>
     /// <returns>status containing a success message, or errors</returns>
@@ -43,7 +43,7 @@ public interface IAccessDatabaseInformation
 
     /// <summary>
     /// This removes a <see cref="DatabaseInformation"/> with the same <see cref="DatabaseInformation.Name"/> as the databaseInfoName.
-    /// If there are no errors it will update the shardingsettings.json file in the application
+    /// If there are no errors it will update the sharding settings file in the application
     /// </summary>
     /// <param databaseInfoName="databaseInfoName">Looks for a <see cref="DatabaseInformation"/> with the <see cref="DatabaseInformation.Name"/> </param>
     /// <returns>status containing a success message, or errors</returns>